feat: figure out blog post display
